# Update substring search to use the Two Way algorithm