… src/libcore/str/mod.rs