@bluss 2015-06-15
cc @BurntSushi, you know substring search very well, cc @Kimundi about the pattern API