@bluss 2015-06-15
I'll show some benchmark results. My workspace for this is temporarily [hosted as a repo](https://github.com/bluss/scratchspace/tree/twoway), if you want to check the quickcheck & benchmark implementations.

`find` and `rfind` are current libcore's `str::find(&str)` and `str::rfind(&str)`. `tw_contains` is two-way forward search, `tw_contains_rev` is two-way reverse search and `tw_paper` is an unmodified transcription of two-way from the paper.

I got a bit pessimistic when I tried out the different bad cases I could find. The performance bug #25483 is the benchmark called aa_100k below. From experience, the main improvemenet in runtime for that benchmark comes from the `byteset` optimization, not due to the TW algorithm itself.

Look at the `short_short` and `short_long` benchmarks for something of a more normal case for substring search.

```
running 45 tests
test bad_naive::find                                    ... bench:       2,553 ns/iter (+/- 112) = 28 MB/s
test bad_naive::rfind                                   ... bench:       2,179 ns/iter (+/- 35) = 33 MB/s
test bad_naive::tw_contains                             ... bench:         952 ns/iter (+/- 260) = 77 MB/s
test bad_naive::tw_contains_rev                         ... bench:         342 ns/iter (+/- 21) = 216 MB/s
test bad_naive::tw_paper                                ... bench:         587 ns/iter (+/- 5) = 126 MB/s
test bad_naive_reversed::find                           ... bench:         691 ns/iter (+/- 8) = 107 MB/s
test bad_naive_reversed::rfind                          ... bench:         690 ns/iter (+/- 133) = 107 MB/s
test bad_naive_reversed::tw_contains                    ... bench:         336 ns/iter (+/- 3) = 220 MB/s
test bad_naive_reversed::tw_contains_rev                ... bench:         735 ns/iter (+/- 21) = 100 MB/s
test bad_naive_reversed::tw_paper                       ... bench:         591 ns/iter (+/- 159) = 125 MB/s
test pathological_aa_100k::find                         ... bench:     999,765 ns/iter (+/- 117,750) = 100 MB/s
test pathological_aa_100k::rfind                        ... bench:   1,001,302 ns/iter (+/- 1,269) = 99 MB/s
test pathological_aa_100k::tw_contains                  ... bench:      16,289 ns/iter (+/- 18) = 6139 MB/s
test pathological_aa_100k::tw_contains_rev              ... bench:       7,196 ns/iter (+/- 197) = 13896 MB/s
test pathological_aa_100k::tw_paper                     ... bench:     706,485 ns/iter (+/- 655) = 141 MB/s
test pathological_aab_100k::find                        ... bench:   4,097,780 ns/iter (+/- 117,012) = 73 MB/s
test pathological_aab_100k::rfind                       ... bench:   3,814,527 ns/iter (+/- 15,899) = 78 MB/s
test pathological_aab_100k::tw_contains                 ... bench:   2,388,729 ns/iter (+/- 11,689) = 125 MB/s
test pathological_aab_100k::tw_contains_rev             ... bench:   3,354,270 ns/iter (+/- 4,309) = 89 MB/s
test pathological_aab_100k::tw_paper                    ... bench:   2,225,404 ns/iter (+/- 93,473) = 134 MB/s
test pathological_two_way_10k::find                     ... bench:     297,304 ns/iter (+/- 24,553) = 100 MB/s
test pathological_two_way_10k::rfind                    ... bench:     297,562 ns/iter (+/- 358) = 100 MB/s
test pathological_two_way_10k::tw_contains              ... bench:     300,044 ns/iter (+/- 447) = 99 MB/s
test pathological_two_way_10k::tw_contains_rev          ... bench:       3,358 ns/iter (+/- 56) = 8933 MB/s
test pathological_two_way_10k::tw_paper                 ... bench:     223,687 ns/iter (+/- 7,375) = 134 MB/s
test pathological_two_way_20k::find                     ... bench:   1,001,881 ns/iter (+/- 117,024) = 99 MB/s
test pathological_two_way_20k::rfind                    ... bench:     997,402 ns/iter (+/- 2,166) = 100 MB/s
test pathological_two_way_20k::tw_contains              ... bench:     919,003 ns/iter (+/- 2,327) = 108 MB/s
test pathological_two_way_20k::tw_contains_rev          ... bench:       5,750 ns/iter (+/- 26) = 17391 MB/s
test pathological_two_way_20k::tw_paper                 ... bench:     727,667 ns/iter (+/- 14,032) = 137 MB/s
test pathological_two_way_20k_reversed::find            ... bench:     738,045 ns/iter (+/- 772) = 81 MB/s
test pathological_two_way_20k_reversed::rfind           ... bench:     706,245 ns/iter (+/- 2,762) = 84 MB/s
test pathological_two_way_20k_reversed::tw_contains     ... bench:       5,829 ns/iter (+/- 36) = 10293 MB/s
test pathological_two_way_20k_reversed::tw_contains_rev ... bench:     295,658 ns/iter (+/- 295) = 202 MB/s
test pathological_two_way_20k_reversed::tw_paper        ... bench:     445,339 ns/iter (+/- 9,289) = 134 MB/s
test short_long::find                                   ... bench:      29,575 ns/iter (+/- 88) = 86 MB/s
test short_long::rfind                                  ... bench:      30,014 ns/iter (+/- 351) = 84 MB/s
test short_long::tw_contains                            ... bench:       7,733 ns/iter (+/- 100) = 329 MB/s
test short_long::tw_contains_rev                        ... bench:       5,910 ns/iter (+/- 41) = 431 MB/s
test short_long::tw_paper                               ... bench:      15,730 ns/iter (+/- 27) = 162 MB/s
test short_short::find                                  ... bench:         212 ns/iter (+/- 1) = 264 MB/s
test short_short::rfind                                 ... bench:         400 ns/iter (+/- 123) = 140 MB/s
test short_short::tw_contains                           ... bench:         164 ns/iter (+/- 0) = 341 MB/s
test short_short::tw_contains_rev                       ... bench:         169 ns/iter (+/- 1) = 331 MB/s
test short_short::tw_paper                              ... bench:         421 ns/iter (+/- 3) = 133 MB/s
```