@bluss 2015-06-16
Thanks for your extensive review @gereeter! I've pushed a commit to address some of this, with some cleanup. I also realized I could move the break for `Reject` until after the fast reject by the byteset. This improves the benchmarks for the forward search uniformly.